requirements.txt
numpy~=1.19.4
scipy~=1.5.4
tensorpack~=0.11
pillow~=8.1.0
opencv-python~=4.5.1.48
mtcnn~=0.1.0
networkx~=2.5
torchvision~=0.8.1
tensorflow~=2.4.1
tensorflow-gpu~=2.4.1